streamlit_app/requirements.txt
# streamlit_app/requirements.txt
streamlit>=1.44.0
httpx>=0.28.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio>=3.4.3
websockets>=11.0.3
redis>=5.0.0

# Data processing and visualization
pandas>=2.0.0
numpy>=1.24.0
altair>=5.0.0
plotly>=5.14.0